refactor(render-svg): split parts.ts into focused modules
No behavior change: snapshot stays byte-identical. Splits XML escapes,
eyes, mouth, and accessories into separate files so the upcoming variation
work can land in narrow diffs.

Spec: docs/superpowers/specs/2026-05-03-pet-visuals-a-plus-b-design.md
