+++ packages/render-svg/src/xml.ts
@@ -1,0 +1,10 @@
+export function escapeXmlText(s: string): string {
+  return s
+    .replace(/&/g, "&amp;")
+    .replace(/</g, "&lt;")
+    .replace(/>/g, "&gt;");
+}
+
+export function escapeXmlAttr(s: string): string {
+  return escapeXmlText(s).replace(/"/g, "&quot;").replace(/'/g, "&apos;");
+}
